Google interview quesiton, RemoveKthElement from last of linkedlist
Used 2 pointers "fast" and "slow" to traverse, since it has to be done in 1 pass:
-
This problem was asked by Google.
Given a singly linked list and an integer k, remove the k last element from the list. k is guaranteed
to be smaller than the length of the list.
The list is very long, so making more than one pass is prohibitively expensive.
Do this in constant space and in one pass.